+++ src/TestSuite/TestCase.php
@@ -50,10 +50,10 @@ abstract class TestCase extends CakeTestCase
         parent::tearDown();
 
         if (LOGS !== TMP) {
-            unlink_recursive(LOGS, 'empty');
+            @unlink_recursive(LOGS, 'empty');
         }
 
-        unlink_recursive(WWW_ROOT . 'vendor', 'empty');
+        @unlink_recursive(WWW_ROOT . 'vendor', 'empty');
         @unlink(WWW_ROOT . 'me_tools');
         @unlink(WWW_ROOT . 'robots.txt');
     }